pyproject.toml
[project]
name = "steam-workshop-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.14.0",
    "bs4>=0.0.2",
    "html2text>=2025.4.15",
    "lxml>=6.0.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.9",
    "sqlalchemy>=2.0.36",
]
